+++ changelog.md
@@ -6,6 +6,7 @@
 - Fixed Entropynnium's detonation projectiles not working with Chance Bullets or Ring of Triggers
 - Fixed Cleansweep's Minesweeper grid labels being visible while the game is paused
 - Fixed null deref caused by trying to disable a non-existent renderer when quick restarting with an active Minesweeper grid
+- Fixed Gradius not working with Backup Gun, Ring of Triggers, Scattershot, Chance Bullets, or Duct Tape
 
 ## 1.29.0 (2026-04-23)